Add files via upload
FlexiFund is a decentralized crowdfunding platform built on blockchain technology. It allows individuals and organizations to create and manage fundraising campaigns with transparency and security. With FlexiFund, users can contribute to projects they believe in using cryptocurrencies, ensuring fast and efficient transactions. The platform leverages smart contracts to automate the fundraising process, providing a flexible and trustworthy solution for crowdfunding initiatives.